==> tasks/201504_stat4701/00-combine-orig-with-pickup-dropoff-roadbed-distances.ipynb <==
import pandas as pd
from datetime import datetime
import os

dboxData = '/Volumes/Hotel/Dropbox/GIS/Data/Municipal/USA/New_York/New_York_City/Taxi/tripData2013/'
d = dboxData
outfolder = '/Users/danielmsheehan/Desktop/data/tables/'
o = outfolder

start = datetime(2013, 1, 1)
end  = datetime(2013, 12, 31)

daterange = pd.date_range(start, end)

for i in daterange:
    z = i.strftime("%Y%m%d")
    m = z[4:6]
    m = m.lstrip('0')
    print z
    print m

    dfo = pd.read_csv(d + 'day/trip_data_'+z+'_m'+m+'.csv')
    dfd = pd.read_csv(d + 'tables/trip_'+z+'_d_roadbed.csv')
    dfp = pd.read_csv(d + 'tables/trip_'+z+'_p_roadbed.csv')

    date = int(z) * 1000000

    dfo['tuid'] = date + dfo.index + 1

    dfd['tuid'] = date + dfd.IN_FID
    dfd['dist_roadbed_dropoff_feet'] = dfd['NEAR_DIST']
    dfd = dfd[['tuid','dist_roadbed_dropoff_feet']]

    dfp['tuid'] = date + dfp.IN_FID
    dfp['dist_roadbed_pickup_feet'] = dfp['NEAR_DIST']
    dfp= dfp[['tuid','dist_roadbed_pickup_feet']]

    df = dfo.merge(dfp, on='tuid', how='left').merge(dfd, on='tuid', how='left')

    resort = ['tuid','trip_distance','pickup_longitude','pickup_latitude','dropoff_longitude','dropoff_latitude','pickup_date','dropoff_date','dist_roadbed_pickup_feet','dist_roadbed_dropoff_feet']

    df = df[resort]

    df.to_csv(o+z+'.csv', index=False)
