==> tasks/201504_stat4701/00-export-building-data.py <==
arcpy.ExportXYv_stats("V:/GIS/projects/gps/tasks/201406_building_bulk_density/data/processing/census.gdb/nycb2010_bulkdens","OBJECTID;CB2010;BoroCode;BoroName;CT2010;BCTCB2010;Shape_Leng;geoid;origblarea;building_block_int_dis_tbl_OBJECTID;building_block_int_dis_tbl_geoid;building_block_int_dis_tbl_origblarea;building_block_int_dis_tbl_SUM_buildingvolume;building_block_int_dis_tbl_Shape_Length;building_block_int_dis_tbl_Shape_Area;building_block_int_dis_tbl_bulkdens;Shape_Length;Shape_Area","COMMA","Z:/Desktop/data/input/building/bldg_dist_height.csv","ADD_FIELD_NAMES")
