==> tasks/201504_stat4701/01-split-xy-for-pickups-and-dropoffs.ipynb <==
import pandas as pd
from datetime import datetime
import os

dboxData = '/Volumes/Hotel/Dropbox/GIS/Data/Municipal/USA/New_York/New_York_City/Taxi/tripData2013/'
d = dboxData
outfolder = '/Users/danielmsheehan/Desktop/data/tables/'
o = outfolder

start = datetime(2013, 1, 1)
end  = datetime(2013, 12, 31)

daterange = pd.date_range(start, end)

for i in daterange:
    z = i.strftime("%Y%m%d")
    m = z[4:6]
    m = m.lstrip('0')
    print z
    print m
    df = pd.read_csv(o+z+'.csv')
    df = df[['tuid','pickup_longitude','pickup_latitude']]
    df['lng'] = df['pickup_longitude']
    df['lat'] = df['pickup_latitude']
    df = df[['tuid','lng','lat']]
    df.to_csv(o+'p/'+z+'.csv', index=False)
    
    df = pd.read_csv(o+z+'.csv')
    df = df[['tuid','dropoff_longitude','dropoff_latitude']]
    df['lng'] = df['dropoff_longitude']
    df['lat'] = df['dropoff_latitude']
    df = df[['tuid','lng','lat']]
    df.to_csv(o+'d/'+z+'.csv', index=False)
