==> tasks/201504_stat4701/02-bring-csv-to-shp.ipynb <==
import re
import pandas as pd
from shapely.geometry import Point, mapping
from fiona import collection
import glob
import os
from datetime import datetime

start = datetime(2013, 1, 19)
end  = datetime(2013, 12, 31)

daterange = pd.date_range(start, end)

for i in daterange:
    z = i.strftime("%Y%m%d")
    print z

    df = pd.read_csv('/Users/danielmsheehan/Desktop/data/tables/p/'+z+'.csv')
    ioSHP = '/Users/danielmsheehan/Desktop/data/processing/p/p_'+z+'.shp'

    lng = 'lng'
    lat = 'lat'

    schema = { 'geometry': 'Point', 'properties': { 'tuid': 'str','lat':'float','lng':'float'} }

    data = df
    with collection(ioSHP, "w", "ESRI Shapefile", schema) as output:
        for index, row in data.iterrows():
            point = Point(row[lng], row[lat])
            output.write({
                'properties': {'tuid': row['tuid'],'lat': row['lat'],'lng': row['lng']},
                'geometry': mapping(point)
            })
    
    df = pd.read_csv('/Users/danielmsheehan/Desktop/data/tables/d/'+z+'.csv')
    ioSHP = '/Users/danielmsheehan/Desktop/data/processing/d/d_'+z+'.shp'

    lng = 'lng'
    lat = 'lat'

    schema = { 'geometry': 'Point', 'properties': { 'tuid': 'str','lat':'float','lng':'float'} }

    data = df
    with collection(ioSHP, "w", "ESRI Shapefile", schema) as output:
        for index, row in data.iterrows():
            point = Point(row[lng], row[lat])
            output.write({
                'properties': {'tuid': row['tuid'],'lat': row['lat'],'lng': row['lng']},
                'geometry': mapping(point)
            })

Not using any of this: ignore or delete later. 

    #ogr2ogr -a_srs EPSG:4326 /Users/danielmsheehan/Desktop/data/processing/p/p_20130101_wgs84_prj.shp /Users/danielmsheehan/Desktop/data/processing/p/p_20130101.shp 
    
     2263
    ogr2ogr -a_srs EPSG:2263 /Users/danielmsheehan/Desktop/data/processing/p/p_20130101_nyspli_prj.shp /Users/danielmsheehan/Desktop/data/processing/p/p_20130101_wgs84_prj.shp
    ogr2ogr -f "ESRI Shapefile" /Users/danielmsheehan/Desktop/data/processing/p/p_20130101_wgs84_prj.shp /Users/danielmsheehan/Desktop/data/processing/p/p_20130101_nyspli_prj.shp -s_srs EPSG:4326 -t_srs EPSG:2263
    
    ogr2ogr -t_srs EPSG:2263 /Users/danielmsheehan/Desktop/data/processing/p/p_20130101_nyspli_prj.shp /Users/danielmsheehan/Desktop/data/processing/p/p_20130101_wgs84_prj.shp
