==> tasks/201504_stat4701/03-csv-xy-intersect.ipynb <==
import glob

csvList = glob.glob('/Users/danielmsheehan/Desktop/data/tables/'+"*.csv")

x = []

for i in csvList:
    i = i.replace('/Users/danielmsheehan/Desktop/data/tables/','').replace('.csv','')
    #print i
    x.append(i)

#print x
z = x[:30]
z = x[30:120]
z = x[112:120]
z = x[120:200]
z = x[200:280]
z = x[280:365]

print z

import glob

csvList = glob.glob('Z:/Desktop/data/tables/'+"*.csv")

x = []

for i in csvList:
    i = i.replace('Z:/Desktop/data/tables\\','').replace('.csv','')
    #print i
    x.append(i)

#print x
z = x[:30]
z = x[30:120]
z = x[112:120]
z = x[120:200]
z = x[200:280]
z = x[280:365]

print z

import arcpy #Doing in ArcGIS is the way to go, much faster, just gotta stick with Shapefiles

for i in z:
    arcpy.Project_management("Z:/Desktop/data/processing/p/p_"+i+".shp","Z:/Desktop/data/processing/p/p_"+i+"_spli.shp","PROJCS['NAD_1983_StatePlane_New_York_Long_Island_FIPS_3104_Feet',GEOGCS['GCS_North_American_1983',DATUM['D_North_American_1983',SPHEROID['GRS_1980',6378137.0,298.257222101]],PRIMEM['Greenwich',0.0],UNIT['Degree',0.0174532925199433]],PROJECTION['Lambert_Conformal_Conic'],PARAMETER['False_Easting',984250.0],PARAMETER['False_Northing',0.0],PARAMETER['Central_Meridian',-74.0],PARAMETER['Standard_Parallel_1',40.66666666666666],PARAMETER['Standard_Parallel_2',41.03333333333333],PARAMETER['Latitude_Of_Origin',40.16666666666666],UNIT['Foot_US',0.3048006096012192]]","WGS_1984_(ITRF00)_To_NAD_1983","GEOGCS['GCS_WGS_1984',DATUM['D_WGS_1984',SPHEROID['WGS_1984',6378137.0,298.257223563]],PRIMEM['Greenwich',0.0],UNIT['Degree',0.0174532925199433],METADATA['World',-180.0,-90.0,180.0,90.0,0.0,0.0174532925199433,0.0,1262]]")
    arcpy.Project_management("Z:/Desktop/data/processing/d/d_"+i+".shp","Z:/Desktop/data/processing/d/d_"+i+"_spli.shp","PROJCS['NAD_1983_StatePlane_New_York_Long_Island_FIPS_3104_Feet',GEOGCS['GCS_North_American_1983',DATUM['D_North_American_1983',SPHEROID['GRS_1980',6378137.0,298.257222101]],PRIMEM['Greenwich',0.0],UNIT['Degree',0.0174532925199433]],PROJECTION['Lambert_Conformal_Conic'],PARAMETER['False_Easting',984250.0],PARAMETER['False_Northing',0.0],PARAMETER['Central_Meridian',-74.0],PARAMETER['Standard_Parallel_1',40.66666666666666],PARAMETER['Standard_Parallel_2',41.03333333333333],PARAMETER['Latitude_Of_Origin',40.16666666666666],UNIT['Foot_US',0.3048006096012192]]","WGS_1984_(ITRF00)_To_NAD_1983","GEOGCS['GCS_WGS_1984',DATUM['D_WGS_1984',SPHEROID['WGS_1984',6378137.0,298.257223563]],PRIMEM['Greenwich',0.0],UNIT['Degree',0.0174532925199433],METADATA['World',-180.0,-90.0,180.0,90.0,0.0,0.0174532925199433,0.0,1262]]")

for i in z: #filename includes wgs84, probably should not as it is state plane
    arcpy.Intersect_analysis("Z:/Desktop/data/processing/p/p_"+i+"_spli.shp #;Z:/Desktop/data/input/census/nycb2010.shp #","Z:/Desktop/data/processing/p/p_"+i+"_wgs84_prj_int_cb2010.shp","ALL","#","INPUT")
    arcpy.Intersect_analysis("Z:/Desktop/data/processing/d/d_"+i+"_spli.shp #;Z:/Desktop/data/input/census/nycb2010.shp #","Z:/Desktop/data/processing/d/d_"+i+"_wgs84_prj_int_cb2010.shp","ALL","#","INPUT")
