==> tasks/201504_stat4701/05-combine-tables.ipynb <==
import pandas as pd
import glob

pj = '/Users/danielmsheehan/Desktop/data/'
#pj = '/Users/danielmsheehan/Dropbox/data_temp/'

csvList = glob.glob(pj+'tables/'+"*.csv")

x = []

for i in csvList:
    i = i.replace(pj+'tables/','').replace('.csv','')
    #print i
    x.append(i)

dfb = pd.read_csv(pj+'input/building/bldg_dist_height.csv').rename(columns=lambda x: x.lower())
dfb = dfb[['geoid','building_block_int_dis_tbl_bulkdens']]
dfb = dfb.fillna(0)  
dfb['geoid'] = dfb['geoid'].astype(str)
print dfb.dtypes

for i in x:
    df = pd.read_csv(pj+'tables/'+i+'.csv')
    
    dfcb = pd.read_csv(pj+'processing/p/p_'+i+'_int_cb2010.csv')
    dfcb = dfcb[['tuid','geoid']]
    dfcb['geoid'] = dfcb['geoid'].astype(str)
    #print dfcb.dtypes
    
    df = df.merge(dfcb, on='tuid', how='left')
    #print df.dtypes
    
    df = df[['tuid','dist_roadbed_pickup_feet','geoid']]
    #print df.dtypes
    df = df.merge(dfb, on='geoid', how='left')
    df.to_csv(pj+'output/tables/p_'+i+'.csv', index=False)
    
    print i + ' is complete for pickups'
    
for i in x:
    df = pd.read_csv(pj+'tables/'+i+'.csv')
    
    dfcb = pd.read_csv(pj+'processing/d/d_'+i+'_int_cb2010.csv')
    dfcb = dfcb[['tuid','geoid']]
    dfcb['geoid'] = dfcb['geoid'].astype(str)
    #print dfcb.dtypes
    
    df = df.merge(dfcb, on='tuid', how='left')  
    #print df.dtypes
    
    df = df[['tuid','dist_roadbed_dropoff_feet','geoid']]
    #print df.dtypes
    df = df.merge(dfb, on='geoid', how='left')
    df.to_csv(pj+'output/tables/d_'+i+'.csv', index=False)
    
    print i + ' is complete for dropoffs'
