==> tasks/201504_stat4701/05-R-plots.Rmd <==
---
title: "GPS Distributed Building Volume Vs. Average Taxi Pickup Distance from Roadbed"
author: "DanielMSheehan"
date: "April 7, 2015"
output: html_document
---

This is an R Markdown document. Markdown is a simple formatting syntax for authoring HTML, PDF, and MS Word documents. For more details on using R Markdown see <http://rmarkdown.rstudio.com>.

When you click the **Knit** button a document will be generated that includes both content as well as the output of any embedded R code chunks within the document. You can embed an R code chunk like this:

```{r}
install.packages("ggplot2")
library(ggplot2)



df <- read.csv('/Users/danielmsheehan/Desktop/data/output/tables/p_20130101_avg_dist.csv')
```

You can also embed plots, for example:

```{r, echo=FALSE}
#df <- data.frame(x = rnorm(5000),y=rnorm(5000))
ggplot(df,aes(x=building_block_int_dis_tbl_bulkdens,y=avg_pickup_dist_feet)) + geom_point(alpha = 0.3)
```

Note that the `echo = FALSE` parameter was added to the code chunk to prevent printing of the R code that generated the plot.
