==> tasks/201504_stat4701/06-concat-master.ipynb <==
import pandas as pd
import glob

pj = '/Users/danielmsheehan/Desktop/data/'
pj = '/Users/danielmsheehan/Dropbox/data/'

csvList = glob.glob(pj+'tables/'+"*.csv") #remove _sm

x = []

for i in csvList:
    i = i.replace(pj+'tables/','').replace('.csv','') #remove _sm
    #print i
    x.append(i)
 
print x

df_list = []
for i in x:
    df = pd.read_csv(pj+'output/tables/d_'+i+'.csv', dtype={'geoid':object})
    df['geoid'] = df['geoid'].astype(str)
    df_list.append(df)
#print df_list    
dfc = pd.concat(df_list)
dfc['geoid'] = dfc['geoid'].astype(str)
dfc['type'] = 'd'

dfc.columns = ['tuid','dist_roadbed','geoid','dist_bldg_hght','type']
print dfc.dtypes
dfc.to_csv(pj+'output/d_2013.csv', index=False)

print len(dfc.index)

%reset

import pandas as pd
import glob

pj = '/Users/danielmsheehan/Desktop/data/'
pj = '/Users/danielmsheehan/Dropbox/data/'

csvList = glob.glob(pj+'tables/'+"*.csv") #remove _sm

x = []

for i in csvList:
    i = i.replace(pj+'tables/','').replace('.csv','') #remove _sm
    #print i
    x.append(i)
 
print x

df_list = []
for i in x:
    df = pd.read_csv(pj+'output/tables/p_'+i+'.csv', dtype={'geoid':object})
    df['geoid'] = df['geoid'].astype(str)
    df_list.append(df)
#print df_list    
dfc = pd.concat(df_list)
dfc['geoid'] = dfc['geoid'].astype(str)
dfc['type'] = 'p'

dfc.columns = ['tuid','dist_roadbed','geoid','dist_bldg_hght','type']
print dfc.dtypes
dfc.to_csv(pj+'output/p_2013.csv', index=False)

print len(dfc.index)

%reset

import pandas as pd
import glob

pj = '/Users/danielmsheehan/Desktop/data/'
pj = '/Users/danielmsheehan/Dropbox/data/'

csvList = glob.glob(pj+'output/'+"*.csv") #remove _sm

x = []

for i in csvList:
    i = i.replace(pj+'output/','').replace('.csv','') #remove _sm
    #print i
    x.append(i)
 
print x

df_list = []
for i in x:
    df = pd.read_csv(pj+'output/'+i+'.csv', dtype={'geoid':object})
    df['geoid'] = df['geoid'].astype(str)
    df_list.append(df)
#print df_list    
dfc = pd.concat(df_list)

dfc.to_csv(pj+'output/all/taxi_2013.csv', index=False)

print dfc.dtypes
print len(dfc.index)

%reset
