==> tasks/201504_stat4701/08-collapse-geoid.ipynb <==
import pandas as pd
import glob

pj = '/Users/danielmsheehan/Desktop/data/'
pj = '/Users/danielmsheehan/Dropbox/data/'

df = pd.read_csv(pj+'output/all/taxi_2013.csv', dtype={'geoid':object})

df['count'] = 1

dfg = df.groupby('geoid').sum()

dfg.to_csv(pj+'output/all/taxi_2013_groupby_geoid_temp.csv')

import pandas as pd

pj = '/Users/danielmsheehan/Desktop/data/'
pj = '/Users/danielmsheehan/Dropbox/data/'
pj = '/Volumes/Hotel/Dropbox/data/'

df = pd.read_csv(pj+'output/all/taxi_2013_groupby_geoid_temp.csv', dtype={'geoid':object})

df['avg_dist_roadbed'] = df['dist_roadbed'] / df['count']

df = df[['geoid','avg_dist_roadbed','count']]

df.to_csv(pj+'output/all/taxi_2013_geoid_avg_dist.csv', index=False)

df
