==> tasks/201504_stat4701/09-stats-to-shp.ipynb <==
import arcpy

arcpy.TableToTable_conversion("W:/data/output/all/taxi_2013_describe_stats.csv","W:/data/output/all","taxi_2013_describe_stats.dbf","#","""geoid "geoid" true true false 15 Text 0 0 ,First,#,W:/data/output/all/taxi_2013_describe_stats.csv,geoid,-1,-1;avgbrdist "avgbrdist" true true false 8 Double 0 0 ,First,#,W:/data/output/all/taxi_2013_describe_stats.csv,avgbrdist,-1,-1;medbrdist "medbrdist" true true false 8 Double 0 0 ,First,#,W:/data/output/all/taxi_2013_describe_stats.csv,medbrdist,-1,-1;stdbrdist "stdbrdist" true true false 8 Double 0 0 ,First,#,W:/data/output/all/taxi_2013_describe_stats.csv,stdbrdist,-1,-1;count "count" true true false 4 Long 0 0 ,First,#,W:/data/output/all/taxi_2013_describe_stats.csv,count,-1,-1""","#")

arcpy.MakeFeatureLayer_management("W:/data/output/all/nycb2010_taxi_2013.shp","nycb2010_taxi_2013_Layer","#","#","FID FID VISIBLE NONE;Shape Shape VISIBLE NONE;CB2010 CB2010 VISIBLE NONE;BoroCode BoroCode VISIBLE NONE;BoroName BoroName VISIBLE NONE;CT2010 CT2010 VISIBLE NONE;BCTCB2010 BCTCB2010 VISIBLE NONE;Shape_Leng Shape_Leng VISIBLE NONE;Shape_Area Shape_Area VISIBLE NONE;geoid geoid VISIBLE NONE")

arcpy.AddJoin_management("nycb2010_taxi_2013_Layer","geoid","W:/data/output/all/taxi_2013_describe_stats.dbf","geoid","KEEP_ALL")

arcpy.FeatureClassToFeatureClass_conversion("nycb2010_taxi_2013_Layer","W:/data/output/all","nycb2010_taxi_2013_stats.shp","#","""CB2010 "CB2010" true true false 4 Text 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,nycb2010_taxi_2013.CB2010,-1,-1;BoroCode "BoroCode" true true false 1 Text 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,nycb2010_taxi_2013.BoroCode,-1,-1;BoroName "BoroName" true true false 32 Text 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,nycb2010_taxi_2013.BoroName,-1,-1;CT2010 "CT2010" true true false 6 Text 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,nycb2010_taxi_2013.CT2010,-1,-1;BCTCB2010 "BCTCB2010" true true false 11 Text 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,nycb2010_taxi_2013.BCTCB2010,-1,-1;Shape_Leng "Shape_Leng" true true false 19 Double 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,nycb2010_taxi_2013.Shape_Leng,-1,-1;Shape_Area "Shape_Area" true true false 19 Double 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,nycb2010_taxi_2013.Shape_Area,-1,-1;geoid "geoid" true true false 254 Text 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,nycb2010_taxi_2013.geoid,-1,-1;geoid_1 "geoid_1" true true false 15 Text 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,taxi_2013_describe_stats.geoid,-1,-1;avgbrdist "avgbrdist" true true false 19 Double 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,taxi_2013_describe_stats.avgbrdist,-1,-1;medbrdist "medbrdist" true true false 19 Double 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,taxi_2013_describe_stats.medbrdist,-1,-1;stdbrdist "stdbrdist" true true false 19 Double 0 0 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,taxi_2013_describe_stats.stdbrdist,-1,-1;count "count" true true false 9 Long 0 9 ,First,#,W:/data/output/all/nycb2010_taxi_2013.shp,taxi_2013_describe_stats.count,-1,-1""","#")

arcpy.AddField_management("W:/data/output/all/nycb2010_taxi_2013_stats.shp","areasqft","FLOAT","#","#","#","#","NULLABLE","NON_REQUIRED","#")

arcpy.CalculateField_management("W:/data/output/all/nycb2010_taxi_2013_stats.shp","areasqft","!shape.area@squarefeet!","PYTHON_9.3","#")


%reset

#intersect buildings and census blocks

arcpy.Intersect_analysis("W:/data/output/all/nycb2010_taxi_2013_stats.shp #;building_0913 #","W:/data/processing/buildings/cb2010_building_0913_int.shp","ALL","#","INPUT")


#dissolve buildings into census blocks

arcpy.Dissolve_management("W:/data/processing/buildings/cb2010_building_0913_int.shp","W:/data/processing/buildings/cb2010_building_0913_int_dis.shp","geoid;areasqft","#","MULTI_PART","DISSOLVE_LINES")

# add/cacle new area
arcpy.AddField_management("W:/data/processing/buildings/cb2010_building_0913_int_dis.shp","bldgareaft","DOUBLE","#","#","#","#","NULLABLE","NON_REQUIRED","#")
arcpy.CalculateField_management("W:/data/processing/buildings/cb2010_building_0913_int_dis.shp","bldgareaft","!shape.area@squarefeet!","PYTHON_9.3","#")

# add/calc pct block area. 
arcpy.AddField_management("W:/data/processing/buildings/cb2010_building_0913_int_dis.shp","pctcbbldg","DOUBLE","#","#","#","#","NULLABLE","NON_REQUIRED","#")
arcpy.CalculateField_management("W:/data/processing/buildings/cb2010_building_0913_int_dis.shp","pctcbbldg","!bldgareaft! / !areasqft!","PYTHON_9.3","#")
