==> tasks/201504_stat4701/12-map-plots.ipynb <==
from IPython.display import HTML 
HTML('<iframe align=left width=50% height=800 frameborder=0 src=https://dms2203.cartodb.com/viz/b68c8b70-e2a2-11e4-bf72-0e4fddd5de28/embed_map allowfullscreen webkitallowfullscreen mozallowfullscreen oallowfullscreen msallowfullscreen></iframe> <iframe align=right width=50% height=800 frameborder=0 src=https://dms2203.cartodb.com/viz/c6fdfff2-e2a2-11e4-9ba1-0e9d821ea90d/embed_map allowfullscreen webkitallowfullscreen mozallowfullscreen oallowfullscreen msallowfullscreen></iframe>')


[Link to Steve's Map](https://www.dropbox.com/s/qkmmlpbevt02dcf/gps_map_v02.png?dl=0)
